requirements.txt
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
et-xmlfile==1.1.0
lxml==4.9.2
numpy==1.25.0
openpyxl==3.1.2
pandas==2.0.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pycparser==2.21
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
pytz==2023.3
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
Wand==0.6.11
